Separates group collect diff and group diff into their own files
